requirements.txt
vllm==0.5.4
opentelemetry-sdk>=1.26.0,<1.27.0
opentelemetry-api>=1.26.0,<1.27.0
opentelemetry-exporter-otlp>=1.26.0,<1.27.0
opentelemetry-semantic-conventions-ai>=0.4.1,<0.5.0
grpcio-tools==1.62.1